diagram: Add more flexibility about times.
At the moment, the PostScript prolog can't cope with outputting other than
one calendar day, but most of the rest of the code is more flexible.
Also we now use DateTime more, so we might one day handle the start and
end of summer time sanely (if the trains do).